requirements.txt
django-redis
django-redis-sessions-fork
librabbitmq
sentry_elastic_nodestore
